CCT 1.0.4-M3 Release notes: Improvements and changes: * Overall optimization for the GUI and calibration modules. ** Calibrations should take ~50% less memory on average and run %10-20 faster. ** The GUI panels should also load and refresh considerably faster as well. * Group velocities used for selection criteria during calibration are now configurable. See the schema folder for examples of the new JSON block and the Parameters tab for the new table display. * Data can now be filtered from use in calibration ** Data selection options: *** From the data tab check the 'used' checkbox or multi-select and right click for a context menu to include/exclude more records at once *** From shape, path, and site plots you can right click a plotted point and exclude the data that was used to generate that plot point (use with care as you can exclude more data than you expect this way, especially in path!) *** From the map you can right click a station or event and include/exclude all records that match either that station or event * Map will now auto (re)center to the data loaded if you tap the 'Map' button while the map is already open (this will also bring the map back to the front if its behind other windows). * Measured and reference event information in the Site correction tab have been consolidated into a single Mw results table. * The combined Mw table also includes the number of measurements used to make the measurement for the measured mw fields. * The Mw results tab now also includes three plots for: ** Reference vs. Measured Mw ** Reference vs. Measured Stress ** Standard deviation of final calibrated measurement by event and frequency. *** Sd plot is color coded by # of station records in the measurement * The measured-mw json output for calibration will now include the reference event if there was one for that evid. * We now remove auto-pick indicator AP picks on waveforms if a human moves the F marker. By moving the F marker this tells the auto-picker that the waveform has been human reviewed and it should not move the F marker during future calibrations. * The waveform plot now also plots the synthetic out past the F marker as a dashed line. This is intended to help a human decide if the pick could be moved further out without affecting the fit too much. * The calibration status update bar has been improved with some additional states showing how far along the calibration is. * If a reference stress value is provided the Site calibration will weight that spectra evenly across all frequencies for that event (default is lower frequencies have higher weight).
Bugfixes:
* Fixed a bug where many overlapping points intersecting with the model line in the Shape tab could make some points stop responding to clicks for plotting purposes.
* The waveform plot should now immediately update and repaint the synthetic when you move the F marker (if it had one).
* Waveform plot won't try to plot a synthetic if the start happens before the expected GV onset for the phase. This criteria was already being used to exclude the synthetic during calibration anyway so no sense in plotting it!
* Workaround implemented for reported hang on Mac 10.13+ caused by a deadlock in the native Apple AWT implementation. The application should no longer occasionally hang on startup.

Known bugs:
* Data selection color changes to indicate used/not-used can be inconsistent until you refresh after excluding/including data on the plots and map. The de/selection does work, however.

+++ calibration-service/calibration-model/src/main/java/gov/llnl/gnem/apps/coda/calibration/model/domain/PeakVelocityMeasurementMetadata.java
@@ -20,7 +20,7 @@ public interface PeakVelocityMeasurementMetadata {
 
     public Long getId();
 
-    public Long getVersion();
+    public Integer getVersion();
 
     public WaveformMetadata getWaveform();
